+++ OUDS/Core/ThemesContract/Sources/_ThemesContract.docc/Tutorial_eiffel/Resources/swift-samples/tutorial-eiffel-2-1.swift
@@ -1,0 +1,11 @@
+import OUDSSwiftUI
+import SwiftUI
+
+// MARK: - Eiffel Theme
+
+class EiffelTheme: OrangeTheme {
+
+    override init() {
+        super.init()
+    }
+}