+++ OUDS/Core/ThemesContract/Sources/_ThemesContract.docc/Tutorial_eiffel/Resources/swift-samples/tutorial-eiffel-3-1.swift
@@ -1,0 +1,18 @@
+import OUDSSwiftUI
+import SwiftUI
+
+// MARK: - Eiffel Theme
+
+class EiffelTheme: OrangeTheme {
+
+    override init() {
+        let colors = EiffelThemeColorProvider()
+
+        super.init(
+            colors: EiffelThemeColorProvider(),
+            button: EiffelThemeButtonComponentTokensProvider(
+                colors: colors),
+            switch: EiffelThemeSwitchComponentTokensProvider(
+                colors: colors))
+    }
+}