+++ OUDS/Core/ThemesContract/Sources/_ThemesContract.docc/Tutorial_eiffel/Resources/swift-samples/tutorial-eiffel-7-1.swift
@@ -1,0 +1,26 @@
+import OUDSSwiftUI
+import SwiftUI
+
+// MARK: - Eiffel Theme
+
+class EiffelTheme: OrangeTheme {
+
+    override init() {
+        let borders = EiffelThemeBorderProvider()
+        let colors = EiffelThemeColorProvider()
+
+        super.init(
+            borders: EiffelThemeBorderProvider(),
+            colors: colors,
+            elevations: EiffelThemeElevationProvider(),
+            spaces: EiffelThemeSpaceProvider(),
+            button: EiffelThemeButtonComponentTokensProvider(
+                colors: colors),
+            checkbox: EiffelThemeCheckboxComponentTokensProvider(
+                borders: borders),
+            radioButton: EiffelThemeRadioButtonComponentTokensProvider(
+                borders: borders),
+            switch: EiffelThemeSwitchComponentTokensProvider(
+                colors: colors))
+    }
+}